fix scheduler manager test
this fix divides scheduling of RSP and scheduler managing test apart.
it helps scheduler manager test in independent test namespace to avoid
issues that existing FTC in kubefed namespace has finalizer because
deployed kubefed control plane set it for other tests.
